removed a whole heap of warnings when compiled on Windows 32 and Windows 64.  Also removed the old ANT-API.
git-svn-id: https://groke.otago.ac.nz/ant/ant@1291 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/directory_iterator_compressor.c
@@ -54,7 +54,7 @@ unsigned long size;
 	one byte at the beginning saying which scheme it should use to decompress.  So, we need
 	to add 2 to the length we allocate and 1 the the length we compress.
 */
-object->compressed = new (std::nothrow) char [size = compressor->space_needed_to_compress(object->length + 1)];		// +1 to include the '\0'
+object->compressed = new (std::nothrow) char [(size_t)(size = compressor->space_needed_to_compress((unsigned long)object->length + 1))];		// +1 to include the '\0'
 if (compressor->compress(object->compressed, &size, object->file, (unsigned long)(object->length + 1)) == NULL)
 	exit(printf("Cannot compress document (name:%s)\n", object->filename));
 object->compressed_length = size;
